+++ graduateProject/admin/order_checkout.php
@@ -4,10 +4,11 @@ include("left.inc.php");
 include("footer.inc.php");
 
 //針對該id做delete動作
-if(isset($_GET['id']) && $_GET['id']!='') {
+if(isset($_GET['id']) && $_GET['id']!='' && csrf_verify($_GET['csrf_token'] ?? '')) {
     $id = mysqli_real_escape_string($conn,$_GET['id']);
     $delete = mysqli_query($conn, "DELETE FROM `user_order` WHERE id = '$id'");
 }
+$token = csrf_token();
 ?>
 
 <div class="rightDiv">
@@ -37,15 +38,15 @@ if(isset($_GET['id']) && $_GET['id']!='') {
                     echo "
                     <tr>
                         <td>".$i++."</td>
-                        <td>".$data['order_name']."</td>
-                        <td>".$data['order_email']."</td>
-                        <td>".$data['order_phone']."</td>
-                        <td>".$data['order_address']."</td>
-                        
-                        <td>".$data['oid']."</td>
-                        <td>".$data['order_user_id']."</td>
-                        <td>".$data['qty']."</td>
-                        <td>&nbsp;<a href='?id=".$data['id']."'>Delete</a>&nbsp;</td>
+                        <td>".h($data['order_name'])."</td>
+                        <td>".h($data['order_email'])."</td>
+                        <td>".h($data['order_phone'])."</td>
+                        <td>".h($data['order_address'])."</td>
+
+                        <td>".h($data['oid'])."</td>
+                        <td>".h($data['order_user_id'])."</td>
+                        <td>".h($data['qty'])."</td>
+                        <td>&nbsp;<a href='?id=".h($data['id'])."&csrf_token=".urlencode($token)."' onclick=\"return confirm('確定要刪除嗎？');\">Delete</a>&nbsp;</td>
                     </tr>
                     ";
                 }
